Reintrocuded calculateEscapeChance into AttemptRunPhase and removed run utils
