+++ test/escape-calculations.test.ts
@@ -6,7 +6,6 @@ import { SpeciesId } from "#enums/species-id";
 import GameManager from "#test/testUtils/gameManager";
 import Phaser from "phaser";
 import { afterEach, beforeAll, beforeEach, describe, expect, it, vi } from "vitest";
-import { calculateEscapeChance } from "#app/utils/run-utils";
 
 describe("Escape chance calculations", () => {
   let phaserGame: Phaser.Game;
@@ -44,7 +43,7 @@ describe("Escape chance calculations", () => {
     commandPhase.handleCommand(Command.RUN, 0);
 
     await game.phaseInterceptor.to(AttemptRunPhase, false);
-
+    const phase = game.scene.phaseManager.getCurrentPhase() as AttemptRunPhase;
     // this sets up an object for multiple attempts. The pokemonSpeedRatio is your speed divided by the enemy speed, the escapeAttempts are the number of escape attempts and the expectedEscapeChance is the chance it should be escaping
     const escapeChances: {
       pokemonSpeedRatio: number;
@@ -89,7 +88,7 @@ describe("Escape chance calculations", () => {
         20,
         escapeChances[i].pokemonSpeedRatio * enemySpeed,
       ]);
-      const chance = calculateEscapeChance(game.scene.currentBattle.escapeAttempts);
+      const chance = phase.calculateEscapeChance(game.scene.currentBattle.escapeAttempts);
       expect(chance).toBe(escapeChances[i].expectedEscapeChance);
     }
   });
@@ -115,7 +114,7 @@ describe("Escape chance calculations", () => {
     commandPhase.handleCommand(Command.RUN, 0);
 
     await game.phaseInterceptor.to(AttemptRunPhase, false);
-
+    const phase = game.scene.phaseManager.getCurrentPhase() as AttemptRunPhase;
     // this sets up an object for multiple attempts. The pokemonSpeedRatio is your speed divided by the enemy speed, the escapeAttempts are the number of escape attempts and the expectedEscapeChance is the chance it should be escaping
     const escapeChances: {
       pokemonSpeedRatio: number;
@@ -168,7 +167,7 @@ describe("Escape chance calculations", () => {
         20,
         escapeChances[i].pokemonSpeedRatio * totalEnemySpeed - playerPokemon[0].stats[5],
       ]);
-      const chance = calculateEscapeChance(game.scene.currentBattle.escapeAttempts);
+      const chance = phase.calculateEscapeChance(game.scene.currentBattle.escapeAttempts);
       // checks to make sure the escape values are the same
       expect(chance).toBe(escapeChances[i].expectedEscapeChance);
       // checks to make sure the sum of the player's speed for all pokemon is equal to the appropriate ratio of the total enemy speed
@@ -192,6 +191,7 @@ describe("Escape chance calculations", () => {
     commandPhase.handleCommand(Command.RUN, 0);
 
     await game.phaseInterceptor.to(AttemptRunPhase, false);
+    const phase = game.scene.phaseManager.getCurrentPhase() as AttemptRunPhase;
 
     // this sets up an object for multiple attempts. The pokemonSpeedRatio is your speed divided by the enemy speed, the escapeAttempts are the number of escape attempts and the expectedEscapeChance is the chance it should be escaping
     const escapeChances: {
@@ -250,7 +250,7 @@ describe("Escape chance calculations", () => {
         20,
         escapeChances[i].pokemonSpeedRatio * enemySpeed,
       ]);
-      const chance = calculateEscapeChance(game.scene.currentBattle.escapeAttempts);
+      const chance = phase.calculateEscapeChance(game.scene.currentBattle.escapeAttempts);
       expect(chance).toBe(escapeChances[i].expectedEscapeChance);
     }
   });
@@ -276,6 +276,7 @@ describe("Escape chance calculations", () => {
     commandPhase.handleCommand(Command.RUN, 0);
 
     await game.phaseInterceptor.to(AttemptRunPhase, false);
+    const phase = game.scene.phaseManager.getCurrentPhase() as AttemptRunPhase;
 
     // this sets up an object for multiple attempts. The pokemonSpeedRatio is your speed divided by the enemy speed, the escapeAttempts are the number of escape attempts and the expectedEscapeChance is the chance it should be escaping
     const escapeChances: {
@@ -341,7 +342,7 @@ describe("Escape chance calculations", () => {
         20,
         escapeChances[i].pokemonSpeedRatio * totalEnemySpeed - playerPokemon[0].stats[5],
       ]);
-      const chance = calculateEscapeChance(game.scene.currentBattle.escapeAttempts);
+      const chance = phase.calculateEscapeChance(game.scene.currentBattle.escapeAttempts);
       // checks to make sure the escape values are the same
       expect(chance).toBe(escapeChances[i].expectedEscapeChance);
       // checks to make sure the sum of the player's speed for all pokemon is equal to the appropriate ratio of the total enemy speed
